ASoC: madera: Fix virtual input mux
The virtual input muxes should have been attached back to the
DSP widget when the preloader was turned into a supply widget,
as you can't have a DAPM route going from a non-supply widget to
a supply widget. ie. Supply widgets can only be the source of a
path with non-supply widgets. This patch makes this change.

Change-Id: I4e1e55857416397c961a7d2c80b016eff7be7404
Signed-off-by: Charles Keepax <ckeepax@opensource.wolfsonmicro.com>