mfd: madera: Add register definitions for Cirrus Logic Madera codecs
This patch adds a header file of register definitions for Cirrus
Logic "Madera" class codecs. These codecs are all based off a common
set of hardware IP so have a common register map (with a few minor
device-to-device variations). These are complex devices with a large
number of features and so have a correspondingly large register set.
The registers.h file has been auto-generated from the hardware register
definitions, stripped down to only registers we need to access from
the driver.

Change-Id: Ib6663cc2182b1bb5de6cbe55637c8920161b1ce9
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>