[COMMON] g2d: Add command list completion
To push a command list to H/W, G2D driver should add some required
information to the command list including:
- initialization of H/W for the new job
- information about the layers involved with the job.
- information about secure layers.
The driver should also satisfies the requirment of H/W. It includes
alignment restriction of the number of commands in the command list.
If the number of commands are not multiple of 8, the driver appends
some redundant commands to satisfy the restriction.

Change-Id: Ia986f95de7df8e96907cac0cf7197f1fc3560db8
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>