[COMMON] soc: samsung: el3_mon: change the way to get VA of _text and _etext
The start and end address of kernel text section were got
from kallsyms_lockup_name which returns the address correspond
to symbol name exported by EXPORT_SYMBOL. So, the address can't
be got if the symbol is not exported.

This patch changes the way to get the start and end addresses of
kernel text section to reading them directly. It is possible
because they have been defined in linker script. The driver
can always get them even if the symbol is not exported.

Change-Id: Ibe3ef81ff5010db55af2db11f07cd1b82e1e6685
Signed-off-by: Junho Choi <junhosj.choi@samsung.com>