g2d: remove conditional compile for Secure DRM
CONFIG_EXYNOS_CONTENT_PATH_PROTECTION is defined when Secure DRM is
ready. Statements that depend on CONFIG_EXYNOS_CONTENT_PATH_PROTECTION
are split by #ifdef for the selective compilation.
But the conditional compilation makes code readers hard to understand
the source code and therefore easy to make mistakes during maintenance
of the source code.
Conditional compilation by CONFIG_EXYNOS_CONTENT_PATH_PROTECTION is
also needed to be removed for G2D of Exynos9820 that does not rely on
the secure world to run both of secure and non-secure G2D jobs.

Change-Id: I097c6ab051f8c3cdefdc6a84a9509d94c2f3149d
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>