[media] vb2: add is_unordered callback for drivers
Explicit synchronization benefits a lot from ordered queues, they fit
better in a pipeline with DRM for example so create a opt-in way for
drivers notify videobuf2 that the queue is unordered.

Drivers don't need implement it if the queue is ordered.

v2:	- improve comments for is_unordered flag (Hans)

v3:	- make it bool (Hans)
	- create vb2_ops_set_unordered() helper

Signed-off-by: Gustavo Padovan <gustavo.padovan@collabora.com>

Change-Id: I519c8576f2fd1b83e60a4da40c169b16b86ef825
Signed-off-by: Janghyuck Kim <janghyuck.kim@samsung.com>