[COMMON] media: mfc: move to the new dma-buf-container
The new, working dma-buf-container is introduced by the commit
ee1d9055d ('dma-buf: introduce dma-buf container'). It still supports
the same API functions but the contexts are slightly changed because
the design of dma-buf-container is changed.

Specifically, dmabuf_container_get_count() now returns -EINVAL instead
of 0 for the dma-bufs that is not a dma-buf-container. This change of
return value comes from the fact that dma-buf-container allows having
no dma-buf by its design although it is impossible to create an empty
dma-buf-container in practice.

Change-Id: I38b7534b2e0bb30e8ec331d019a089736d847dbe
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>