[9610] bts: reduce peak bandwidth of MFC/DPU for FHD480fps
Currently, the frame drop occurring when full HD 480fps
encoding because the bandwidth from the MFC and DPU. The
instantaneous high bandwidth of the MFC and DPU increase
the latency of CPU. So, the priority and peak banwidth of
MFC and DPU were limited on the full HD 480fps encoding
scenario. These limitation does not hurt the MFC and DPU
performance on a corresponding scenario.

Change-Id: I61a4859d3913fc8015e47aa68567b437e30566fb
Signed-off-by: Taekki Kim <taekki.kim@samsung.com>