[COMMON] fimc-is2: Flush the old H/W frames for shot done sequence
If there are the old frames in H/W frame manager,
the matched frames in subdev frame manager will be flushed to user as NDONE.
When user gets the shot data of these frames, every kernel virtual address for the frames is unmapped.
To avoid accessing invalid kernel virtual address in H/W frames,
they should also be flushed in shot done sequence.

Change-Id: I2e1262cac42a31fd7c3567885e0afbc82790ab83
Signed-off-by: Sanghoon Lee <shoon114.lee@samsung.com>