[COMMON] fimc-is2: Added new data in shot_ext
UVSP in MCSC needs the value of LSC_CENTER from CGRAS in 3AA and RADIAL_BIQUAD value in 3AA.
Driver calls the get_cal_info function to DDK, DDK fills the values in cal_info.
These values are need to be sync with frame. So the new data was added in shot_ext.

PR JIRA ID: CPR-39

Change-Id: I44e57317a30865733349f037298e49c69dd81fcd
Signed-off-by: Sunmi Lee <carrotsm.lee@samsung.com>