[9610] wlbt: no sable genetion on wifi disconnect
We decided to disable wireless disconnect trigger for now as a quick
workaround and not generate sable on this trigger. On rapid wifi on/off
we need to serialize tar generation and limit number of tar files we keep
which requires more thought and testing which will be done in a separate
bug.

Change-Id: Ia42e5b870404507c8544c17e600f14b4f3615b1f
SCSC-Bug-Id: SSB-43178
Signed-off-by: Purnendu Kapadia <p.kapadia@samsung.com>