[RAMEN9610-10029][COMMON] media: mfc: fix the WAIT state for decoder DRC
In case of UTC, it calls g_fmt after streamoff so can't out from
WAIT_DPB_FLUSH state. This modified that it can wait both g_fmt and
streamoff regardless of the order.
When DRC is detected the wait_state set to (WAIT_G_FMT | WAIT_STOP).
If the g_fmt is called WAIT_G_FMT is cleared and
if the streamoff is called WAIT_STOP is cleared.

Change-Id: I90326d191de2ba164060673e8b124227fda3a397
Signed-off-by: Sunyoung Kang <sy0816.kang@samsung.com>